エラーハンドリングの実装

+++ src/UseCases/Task/DeleteTaskUseCase.go
@@ -1,10 +1,10 @@
 package UseCases
 
 import (
-	"fmt"
 	Domains "go-ddd-rest-api-sample/src/Domains/Task"
 	Requests "go-ddd-rest-api-sample/src/Presentations/Requests/Task"
 	"go-ddd-rest-api-sample/src/Shared"
+	"go-ddd-rest-api-sample/src/Shared/Errors"
 )
 
 type DeleteTaskUseCase struct {
@@ -25,8 +25,16 @@ func (u *DeleteTaskUseCase) Execute(request Requests.DeleteTaskRequest) error {
 		return err
 	}
 
+	task, err := u.taskRepository.GetTask(taskId)
+	if err != nil {
+		return err
+	}
+	if task == nil {
+		return Errors.New("001-001", "指定されたタスクが存在しません")
+	}
+
 	if err := u.taskRepository.DeleteTask(taskId); err != nil {
-		return fmt.Errorf("タスクの削除に失敗しました")
+		return err
 	}
 	return nil
 }
